Переработка test_put_v1_account test_put_v1_account_token
