--- tests/functional/post_v1_account/test_put_v1_account_token.py
@@ -1,74 +1,0 @@
-from json import loads
-
-from dm_api_account.apis.account_api import AccountApi
-from dm_api_account.apis.login_api import LoginApi
-from api_mailhog.apis.mailhog_api import MailhogApi
-from utils import utils
-from restclient.configuration import Configuration as MailhogConfiguration
-from restclient.configuration import Configuration as DmApiConfiguration
-
-
-def test_put_v1_account_token():
-    # регистрация пользователя
-    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
-    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
-
-    account_api = AccountApi(configuration=dm_api_configuration)
-    login_api = LoginApi(configuration=dm_api_configuration)
-    mailhog_api = MailhogApi(configuration=mailhog_configuration)
-    login = utils.generate_login()
-    password = '123456789'
-    email = f'{login}@mail.ru'
-    json_data = {
-        'login': login,
-        'email': email,
-        'password': password,
-    }
-
-    response = account_api.post_v1_account(json_data=json_data)
-    status_code = response.status_code
-    assert status_code == 201, f"User is not created {response.json()}"
-
-    # получить письма из почтового сервера
-
-    response = mailhog_api.get_api_v2_messages()
-    print(f'{response.status_code}')
-    status_code = response.status_code
-    assert status_code == 200, f"No letters are received. Status code is {status_code}"
-
-    # Получить токен
-    token = get_activation_token_by_login(login, response)
-    assert token is not None, f"No token for user {login}"
-
-    # неверный токен
-    wrong_token = utils.generate_random_string(len(token))  # try to use token of the same length
-    # активация пользователя c неверным токеном
-    response = account_api.put_v1_account_token(token=wrong_token)
-    status_code = response.status_code
-    assert status_code == 400, f"User is activated {response.json()}"
-
-    # TODO add parametrized tests with empty token and token belong to another user
-
-    # авторизоваться
-    json_data = {
-        'login': login,
-        'password': password,
-        'rememberMe': True,
-    }
-    response = login_api.post_v1_account_login(json_data=json_data)
-    status_code = response.status_code
-    assert status_code == 403, f"User CAN authorise {response.json()}"
-
-
-def get_activation_token_by_login(
-        login,
-        response
-):
-    token = None
-    for item in response.json()['items']:
-        user_data = loads(item['Content']['Body'])
-        user_login = user_data['Login']
-        if user_login == login:
-            token = user_data['ConfirmationLinkUrl'].split('/')[-1]
-
-    return token
