+++ tests/functional/put_v1_account/test_put_v1_account.py
@@ -1,25 +1,14 @@
-import structlog
 
 from helpers.account_helper import AccountHelper
-from utils import utils
-from restclient.configuration import Configuration as DmApiConfiguration
-from restclient.configuration import Configuration as MailhogConfiguration
 from services.api_mailhog import MailHogApi
 from services.dm_api_account import DMApiAccount
+from utils import utils
+from restclient.configuration import Configuration as MailhogConfiguration
+from restclient.configuration import Configuration as DmApiConfiguration
 
-structlog.configure(
-    processors=[
-        structlog.processors.JSONRenderer(
-            indent=4,
-            ensure_ascii=True,
-            sort_keys=True
-        )
-    ]
-)
-
-
-def test_post_v1_account():
+def test_put_v1_account():
     # регистрация пользователя
+
     mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
     dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
 
@@ -37,3 +26,11 @@ def test_post_v1_account():
     # авторизоваться
     response = account_helper.user_login(login=login, password=password)
     assert response.status_code == 200, f"User cannot authorise {response.json()}"
+
+    # смена email
+    new_email = f'{login}_NEW@mail.ru'
+    account_helper.put_v1_account_email(login=login, password=password, new_email=new_email)
+
+    # авторизоваться после смены пароля
+    response = account_helper.user_login(login=login, password=password)
+    assert response.status_code == 403, f"User can authorise after email has been altered {response.json()}"
